examples/md_to_pptx_agent/pyproject.toml
[project]
name = "md-to-pptx-agent"
version = "0.1.0"
description = "Markdown to PPTX conversion agent example"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-openai>=0.0.8",
    "langchain-core>=0.3.31",
    "middleman-ai",
    "python-dotenv>=1.0.1",
]

[tool.uv]
[tool.uv.sources]
middleman-ai = { path = "../../", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "src/md_to_pptx_agent",
]
